requirements.txt
# Flask Framework
Flask==0.10.1
Flask-SQLAlchemy==2.0

# Flask Packages
Flask-Social==1.6.2
Flask-WTF==0.10.2

#Assets
Flask-Assets==0.11
jsmin==2.2.0
cssmin==0.2.0

#OAuth
facebook-sdk==1.0

# Development tools
pytest==2.6.3

